remove duplicate test file and update name for test file with analytics
